working with card schema

+++ src/schema/card.schema.ts
@@ -1,7 +1,8 @@
+import { Double } from "mongodb";
 import * as mongoose from 'mongoose';
 
 export const CardSchema = new mongoose.Schema({
     uuid: String,
     owner: String,
-    balance: Double,
+    balance: Integer
 });
